print name of DB causing failure condition
nodeos actually uses two chainbase DBs: state & reversible. When there is an error on one of these DBs (like that it’s dirty, or it fails the environment check) it’s not clear which DB is the culprit. Add the name of the DB in the exception so it is clear.